integrate searchbar component
