+++ components/views/Home.js
@@ -1,43 +1,47 @@
 import React from "react";
-import { Button, View, Text, StyleSheet, TextInput, Pressable } from "react-native";
-import {LinearGradient} from 'expo-linear-gradient';
-import Searchbar from "./ui/Searchbar";
+import {
+    Button,
+    View,
+    Text,
+    StyleSheet,
+    TextInput,
+    Pressable,
+} from "react-native";
+import { LinearGradient } from "expo-linear-gradient";
+import SearchBar from "../ui/SearchBar"
 
 export default function Home({ navigation }) {
     return (
         <View style={styles.container}>
-            
             <LinearGradient
-                colors={["#FF6666", "#FCAEAE", '#FFEADD']}
+                colors={["#FF6666", "#FCAEAE", "#FFEADD"]}
                 start={{ x: 0, y: 1 }}
                 end={{ x: 1, y: 0 }}
                 style={styles.gradient}
             >
-
-                <View style = {styles.header}>
-                    <Text style = {styles.headerText}>Parking Finder</Text>
+                <View style={styles.header}>
+                    <Text style={styles.headerText}>Parking Finder</Text>
                 </View>
 
-                <View style ={styles.main}>
+                <View style={styles.main}>
+                    <Text style={styles.description}>
+                        Enter an address to find parking nearby
+                    </Text>
 
-                <Text style = {styles.description}>Enter an address to find parking nearby</Text>
+                    {/* <View style={styles.search}>
+                        <TextInput
+                            style={styles.input}
+                            placeholder="enter address"
+                        ></TextInput>
+                        <Pressable
+                            style={styles.button}
+                            onPress={() => navigation.navigate("Map")}
+                        >
+                            <Text style={styles.buttonText}>Submit</Text>
+                        </Pressable>
+                    </View> */}
 
-                <View style={styles.search}>
-
-<TextInput style ={styles.input} placeholder="enter address"></TextInput>
-{/* <Button
-    style = {styles.button}
-    title="Submit"
-    onPress={() => navigation.navigate("Map")}
-/> */}
-<Pressable
-    style = {styles.button}
-    onPress={() => navigation.navigate("Map")}
->
-    <Text style = {styles.buttonText}>Submit</Text>
-</Pressable>
-</View>
-
+                    <SearchBar navigation={navigation}/>
                 </View>
             </LinearGradient>
         </View>
@@ -53,34 +57,33 @@ const styles = StyleSheet.create({
     main: {
         flex: 1,
         alignItems: "center",
-        marginTop: "50%"
+        marginTop: "50%",
     },
     header: {
         marginTop: "10%",
         fontSize: 20,
-
     },
-    headerText : {
-        fontSize: 40
+    headerText: {
+        fontSize: 40,
     },
     description: {
-        marginBottom: 20
+        marginBottom: 20,
     },
     search: {
-        flexDirection: "row"
+        flexDirection: "row",
     },
     button: {
         flexDirection: "column",
         justifyContent: "center",
         borderWidth: 1,
-        width: '15%',
-        borderColor: 'black',
+        width: "15%",
+        borderColor: "black",
         borderRadius: 20,
-        marginLeft: 5
+        marginLeft: 5,
     },
     buttonText: {
-        textAlign: 'center',
-        color: 'white'
+        textAlign: "center",
+        color: "white",
     },
     input: {
         height: 40,
@@ -89,7 +92,7 @@ const styles = StyleSheet.create({
         borderColor: "black",
         borderRadius: "20px",
         backgroundColor: "white",
-        textAlign: "center"
+        textAlign: "center",
     },
     gradient: {
         flex: 1,
